functions/api_monolith/requirements.txt
aws-lambda-powertools==2.9.1
PyNaCl==1.5.0
pydantic==1.10.7
